+++ test/string_extensions_test.dart
@@ -6,86 +6,86 @@ void main() {
   test(
     'String words are correctly counted',
     () {
-      var string = 'Hello world ! dear 4asd';
+      String? string = 'Hello world ! dear 4asd';
       expect(string.countWords(), 4);
     },
   );
   test(
     'Returns the read time of the string',
     () {
-      var string = 'Hello dear friend how you doing ?';
+      String? string = 'Hello dear friend how you doing ?';
       expect(string.readTime(), 3);
     },
   );
   test(
     'String should be capitalized',
     () {
-      var string = 'hAckErrR';
+      String? string = 'hAckErrR';
       expect(string.capitalize(), 'Hackerrr');
     },
   );
   test(
     'String should be title cased',
     () {
-      var string = 'Hello dear friend how you doing ?';
+      String? string = 'Hello dear friend how you doing ?';
       expect(string.toTitleCase(), 'Hello Dear Friend How You Doing ?');
     },
   );
   test(
     'Removes all numbers from the string',
     () {
-      var string = '234112731235es4234123e5523nt1is';
+      String? string = '234112731235es4234123e5523nt1is';
       expect(string.removeNumbers(), 'esentis');
     },
   );
   test(
     'Removes everything but latin characters',
     () {
-      var string = '4*%^55/es4e55?:"///23nt1is';
+      String? string = '4*%^55/es4e55?:"///23nt1is';
       expect(string.onlyLatin(), 'esentis');
 
-      var string2 = '4*%^55/es4e55?:"///23nt1isαβδα';
+      String? string2 = '4*%^55/es4e55?:"///23nt1isαβδα';
       expect(string2.onlyLatin(), 'esentis');
 
-      var string3 = '4*%^55/es4e55?:"///23nt1isαβδα ma455 ma2231n';
+      String? string3 = '4*%^55/es4e55?:"///23nt1isαβδα ma455 ma2231n';
       expect(string3.onlyLatin(), 'esentis ma man');
     },
   );
   test(
     'Removes everything but greek characters',
     () {
-      var string = '4*%^σοφ55ία/es4e55?:"///23nt1is';
+      String? string = '4*%^σοφ55ία/es4e55?:"///23nt1is';
       expect(string.onlyGreek(), 'σοφία');
 
-      var string2 = '4*%^σοφ55ία/es4e55?:"///23nt1is αγάπη';
+      String? string2 = '4*%^σοφ55ία/es4e55?:"///23nt1is αγάπη';
       expect(string2.onlyGreek(), 'σοφία αγάπη');
     },
   );
   test(
     'String should be camel cased',
     () {
-      var string = 'Find max of array';
+      String? string = 'Find max of array';
       expect(string.toCamelCase(), 'findMaxOfArray');
     },
   );
   test(
     'Removes everything but numbers',
     () {
-      var string = '4*%^55/es4e5523nt1is';
+      String? string = '4*%^55/es4e5523nt1is';
       expect(string.onlyNumbers(), '455455231');
     },
   );
   test(
     'Removes all letters',
     () {
-      var string = '1244e*s*4e*5523n*t*1i*s';
+      String? string = '1244e*s*4e*5523n*t*1i*s';
       expect(string.removeLetters(), '1244**4*5523**1*');
     },
   );
   test(
     'Correctly finds character occurances',
     () {
-      var string = 'esssssentis';
+      String? string = 'esssssentis';
       expect(string.charOccurences(), [
         {'e': 2},
         {'i': 1},
@@ -97,213 +97,243 @@ void main() {
   test(
     'Correctly finds most frequent character',
     () {
-      var string = 'esssssentis';
+      String? string = 'esssssentis';
       expect(string.mostFrequent(), 's');
     },
   );
   test(
     'Reverses a string',
     () {
-      var string = 'Hello World';
+      String? string = 'Hello World';
       expect(string.reverse(), 'dlroW olleH');
     },
   );
   test(
     'Checks if a string is a valid mail',
     () {
-      var string1 = 'esentis@esentis.com';
+      String? string1 = 'esentis@esentis.com';
       expect(string1.isMail(), true);
 
-      var string2 = 'esentisesentis.com';
+      String? string2 = 'esentisesentis.com';
       expect(string2.isMail(), false);
     },
   );
   test(
     'Checks if a string is a valid Ipv4',
     () {
-      var string1 = '192.168.1.14';
+      String? string1 = '192.168.1.14';
       expect(string1.isIpv4(), true);
 
-      var string2 = '192.168.1.14.5223';
+      String? string2 = '192.168.1.14.5223';
       expect(string2.isIpv4(), false);
     },
   );
   test(
     'Checks if a string is a valid Ipv6',
     () {
-      var string1 = '2001:0db8:85a3:0000:0000:8a2e:0370:7334';
+      String? string1 = '2001:0db8:85a3:0000:0000:8a2e:0370:7334';
       expect(string1.isIpv6(), true);
 
-      var string2 = '192.168.1.14';
+      String? string2 = '192.168.1.14';
       expect(string2.isIpv6(), false);
     },
   );
   test(
     'Checks if a string is a number',
     () {
-      var string1 = '45';
+      String? string1 = '45';
       expect(string1.isNumber(), true);
 
-      var string2 = '45s';
+      String? string2 = '45s';
       expect(string2.isNumber(), false);
 
-      var string3 = '45.5';
+      String? string3 = '45.5';
       expect(string3.isNumber(), true);
     },
   );
   test(
     'Checks if a string is contained in a give Iterable of Strings',
     () {
-      var string1 = '45';
-      var array1 = ['45', '66'];
+      String? string1 = '45';
+      List<String?> array1 = ['45', '66'];
       expect(string1.isIn(array1), true);
 
-      var string2 = 'test';
-      var array2 = ['45', '66', 'ffff', '2123'];
+      String? string2 = 'test';
+      List<String?> array2 = ['45', '66', 'ffff', '2123'];
       expect(string2.isIn(array2), false);
 
-      var string3 = 'test';
-      var array3 = {'45', '66', 'ffff', '2123', 'test'};
+      String? string3 = 'test';
+      List<String?> array3 = ['45', '66', 'ffff', '2123', 'test'];
       expect(string3.isIn(array3), true);
     },
   );
   test(
     'Checks if the string has only latin characters',
     () {
-      var string1 = 'hello friend Γιωργο';
+      String? string1 = 'hello friend Γιωργο';
       expect(string1.isLatin(), false);
 
-      var string2 = 'test world';
+      String? string2 = 'test world';
       expect(string2.isLatin(), true);
 
-      var string3 = 'testworld';
+      String? string3 = 'testworld';
       expect(string3.isLatin(), true);
     },
   );
   test(
     'Checks if the string has only greek characters',
     () {
-      var string1 = 'hello friend Γιωργο';
+      String? string1 = 'hello friend Γιωργο';
       expect(string1.isGreek(), false);
 
-      var string2 = 'Τα αγαθα κοποις κτωνται';
+      String? string2 = 'Τα αγαθα κοποις κτωνται';
       expect(string2.isGreek(), true);
 
-      var string3 = 'Τα αγαθά κόποις κτώνται';
+      String? string3 = 'Τα αγαθά κόποις κτώνται';
       expect(string3.isGreek(), true);
     },
   );
   test(
     'Do something if string is empty',
     () {
-      var string1 = '45';
+      String? string1 = '45';
       expect(string1.ifEmpty(() => 'empty'), '45');
 
-      var string2 = '';
+      String? string2 = '';
       expect(string2.ifEmpty(() => 'empty'), 'empty');
+    },
+  );
+  test(
+    'Do something if string is null',
+    () {
+      String? string1;
+      expect(string1.ifNull(() => 'is null'), 'is null');
+
+      String? string2 = 'esentis';
+      expect(string2.ifNull(() => 'is null'), 'esentis');
+    },
+  );
+  test(
+    'Provide default value if the string is null',
+    () {
+      String? string1;
+      expect(string1.defaultValue('no null please'), 'no null please');
+
+      String? string2 = 'esentis';
+      expect(string2.defaultValue('no null please'), 'esentis');
+    },
+  );
+  test(
+    'Check if a string is null',
+    () {
+      String? string1;
+      expect(string1.isNull(), true);
+
+      String? string2 = 'esentis';
+      expect(string2.defaultValue('no null please'), 'esentis');
     },
   );
   test(
     'Repeats the string X times',
     () {
-      var string1 = 'foo';
+      String? string1 = 'foo';
       expect(string1.repeat(0), 'foo');
 
-      var string2 = 'foo';
+      String? string2 = 'foo';
       expect(string2.repeat(3), 'foofoofoo');
     },
   );
   test(
     'Returns first character of the string',
     () {
-      var string = '192.168.1.14';
+      String? string = '192.168.1.14';
       expect(string.first(), '1');
     },
   );
   test(
     'Returns last character of the string',
     () {
-      var string = '192.168.1.14';
+      String? string = '192.168.1.14';
       expect(string.last(), '4');
     },
   );
   test(
     'String should be slug cased',
     () {
-      var string = 'hello world this is esentis';
+      String? string = 'hello world this is esentis';
       expect(string.toSlug(), 'hello_world_this_is_esentis');
     },
   );
   test(
     'Converts the string to num',
     () {
-      var string = '5';
+      String? string = '5';
       expect(string.toNum(), 5);
 
-      var string2 = '5.0';
+      String? string2 = '5.0';
       expect(string2.toNum(), 5.0);
 
-      var string3 = '5.0f';
+      String? string3 = '5.0f';
       expect(string3.toNum(), null);
 
-      var string5 = 'esentis';
+      String? string5 = 'esentis';
       expect(string5.toNum(), null);
     },
   );
   test(
     'Converts the string to int',
     () {
-      var string = '5';
+      String? string = '5';
       expect(string.toInt(), 5);
 
-      var string2 = '5.4';
+      String? string2 = '5.4';
       expect(string2.toInt(), 5);
 
-      var string3 = '5.0f';
+      String? string3 = '5.0f';
       expect(string3.toInt(), null);
 
-      var string5 = 'esentis';
+      String? string5 = 'esentis';
       expect(string5.toInt(), null);
 
-      var string6 = '5.6';
+      String? string6 = '5.6';
       expect(string6.toInt(), 5);
     },
   );
   test(
     'Converts the string to double',
     () {
-      var string = '5';
+      String? string = '5';
       expect(string.toDouble(), 5.0);
 
-      var string2 = '5.0';
+      String? string2 = '5.0';
       expect(string2.toDouble(), 5.0);
 
-      var string3 = '5.0f';
+      String? string3 = '5.0f';
       expect(string3.toDouble(), null);
 
-      var string5 = 'esentis';
+      String? string5 = 'esentis';
       expect(string5.toDouble(), null);
     },
   );
   test(
     'Normalizes greek word',
     () {
-      var string = 'Αριστοτέλης';
+      String? string = 'Αριστοτέλης';
       expect(string.replaceGreek(), 'aristotelis');
     },
   );
   test(
     'Correctly finds given pattern occurences',
     () {
-      var string = 'abracadabra';
-      var pattern = 'abr';
+      String? string = 'abracadabra';
+      String? pattern = 'abr';
       expect(string.findPattern(pattern: pattern), [0, 7]);
     },
   );
   test(
     'Transforms string to string array',
     () {
-      var string = 'abracadabra';
+      String? string = 'abracadabra';
       expect(string.toArray(),
           ['a', 'b', 'r', 'a', 'c', 'a', 'd', 'a', 'b', 'r', 'a']);
     },
@@ -311,14 +341,14 @@ void main() {
   test(
     "Counts a specific character's occurences in a string",
     () {
-      var string = 'abracadabra';
+      String? string = 'abracadabra';
       expect(string.charCount('a'), 5);
     },
   );
   test(
     'Strips all html code',
     () {
-      var string =
+      String? string =
           '<script>Hacky hacky.</script> <p>Here is some text. <span class="bold">This is bold.</span></p>';
       expect(
           string.stripHtml(), 'Hacky hacky. Here is some text. This is bold.');
@@ -327,36 +357,37 @@ void main() {
   test(
     'Checks if string is strong password',
     () {
-      var string1 = 'qwerty';
+      String? string1 = 'qwerty';
       expect(string1.isStrongPassword(), false);
 
-      var string2 = 'Is1!thisStrong';
+      String? string2 = 'Is1!thisStrong';
       expect(string2.isStrongPassword(), true);
     },
   );
   test(
     'Checks if string is a valid Guid',
     () {
-      var string1 = '6d64-4396-8547-1ec1b86e081e';
+      String? string1 = '6d64-4396-8547-1ec1b86e081e';
       expect(string1.isGuid(), false);
 
-      var string2 = '887b7923-6d64-4396-8547-1ec1b86e081e';
+      String? string2 = '887b7923-6d64-4396-8547-1ec1b86e081e';
       expect(string2.isGuid(), true);
     },
   );
   test(
     'Squeezes a string to a specific character',
     () {
-      var string1 = 'employee';
+      String? string1 = 'employee';
       expect(string1.squeeze('e'), 'employe');
 
-      var string2 = '8recommmmmmendation';
+      String? string2 = '8recommmmmmendation';
       expect(string2.squeeze('m'), '8recomendation');
 
-      var string3 = '8recommmmmmendation';
+      String? string3 = '8recommmmmmendation';
       expect(string3.squeeze('8'), '8recommmmmmendation');
 
-      var string4 = '8recommmmmmendation                                     ';
+      String? string4 =
+          '8recommmmmmendation                                     ';
       expect(string4.squeeze(' '), '8recommmmmmendation ');
     },
   );
@@ -364,51 +395,51 @@ void main() {
   test(
     'Checks if string consisted of same characters',
     () {
-      var string1 = 'employee';
+      String? string1 = 'employee';
       expect(string1.hasSameCharacters(), false);
 
-      var string2 = '1111111';
+      String? string2 = '1111111';
       expect(string2.hasSameCharacters(), true);
 
-      var string3 = '8recommmmmmendation';
+      String? string3 = '8recommmmmmendation';
       expect(string3.hasSameCharacters(), false);
 
-      var string4 = 'kkkKKKKkkkKKkKkkkkKKK';
+      String? string4 = 'kkkKKKKkkkKKkKkkkkKKK';
       expect(string4.hasSameCharacters(), true);
     },
   );
   test(
     'Gets the Levenshtein distance of two strings',
     () {
-      var string1 = 'employee';
+      String? string1 = 'employee';
       expect(string1.getLevenshtein('employee'), 0);
 
-      var string2 = '1111111';
+      String? string2 = '1111111';
       expect(string2.getLevenshtein('11111112'), 1);
 
-      var string3 = '8recommmmmmendation';
+      String? string3 = '8recommmmmmendation';
       expect(string3.getLevenshtein('8recoammgnmendation'), 3);
 
-      var string4 = 'dffasdasd';
+      String? string4 = 'dffasdasd';
       expect(string4.getLevenshtein('employee'), 9);
     },
   );
   test(
     'Checks if a string is valid URL',
     () {
-      var string1 = 'employee';
+      String? string1 = 'employee';
       expect(string1.isUrl(), false);
 
-      var string2 = '1111111.com';
+      String? string2 = '1111111.com';
       expect(string2.isUrl(), true);
 
-      var string3 = 'http://8recommmmmmendation';
+      String? string3 = 'http://8recommmmmmendation';
       expect(string3.isUrl(), false);
 
-      var string4 = 'www.google.com';
+      String? string4 = 'www.google.com';
       expect(string4.isUrl(), true);
 
-      var string5 = 'ftp://www.google.com';
+      String? string5 = 'ftp://www.google.com';
       expect(string5.isUrl(), true);
     },
   );
@@ -416,36 +447,36 @@ void main() {
   test(
     'Checks if a string is valid Date format',
     () {
-      var string1 = 'employee';
+      String? string1 = 'employee';
       expect(string1.isDate(), false);
 
-      var string2 = '02-02-2020';
+      String? string2 = '02-02-2020';
       expect(string2.isDate(), true);
 
-      var string3 = '2020-04-06';
+      String? string3 = '2020-04-06';
       expect(string3.isDate(), true);
 
-      var string4 = '02-13-2020';
+      String? string4 = '02-13-2020';
       expect(string4.isDate(), false);
 
-      var string5 = '20120227T132700';
+      String? string5 = '20120227T132700';
       expect(string5.isDate(), true);
     },
   );
   test(
     'Formats the text with a given mask.',
     () {
-      var string1 = 'aaaaaaaabbbbccccddddeeeeeeeeeeee';
-      var mask1 = 'Hello ########-#A###-####-####-############ Oww';
+      String? string1 = 'aaaaaaaabbbbccccddddeeeeeeeeeeee';
+      String? mask1 = 'Hello ########-#A###-####-####-############ Oww';
       expect(string1.formatWithMask(mask1),
           'Hello aaaaaaaa-bAbbb-cccc-dddd-eeeeeeeeeeee Oww');
 
-      var string2 = 'esenti';
-      var mask2 = 'Hello #######';
+      String? string2 = 'esenti';
+      String? mask2 = 'Hello #######';
       expect(string2.formatWithMask(mask2), 'Hello esenti');
 
-      var string3 = 'esentisgreece';
-      var mask3 = 'Hello ####### you are from ######';
+      String? string3 = 'esentisgreece';
+      String? mask3 = 'Hello ####### you are from ######';
       expect(
           string3.formatWithMask(mask3), 'Hello esentis you are from greece');
     },
@@ -453,12 +484,102 @@ void main() {
   test(
     'String operator tests',
     () {
-      var string1 = 'three';
-      var string2 = 'two';
+      String? string1 = 'three';
+      String? string2 = 'two';
       expect(string1 > string2, true);
       expect(string1 < string2, false);
       expect(string1 >= string2, true);
       expect(string1 <= string2, false);
+    },
+  );
+  test(
+    'Removes the first n characters of the string',
+    () {
+      String? string1 = 'three';
+      expect(string1.removeFirst(3), 'ee');
+      expect(string1.removeFirst(0), 'three');
+      expect(string1.removeFirst(-5), 'three');
+      expect(string1.removeFirst(6), '');
+    },
+  );
+  test(
+    'Removes the last n characters of the string',
+    () {
+      String? string1 = 'three';
+      expect(string1.removeLast(3), 'th');
+      expect(string1.removeLast(0), 'three');
+      expect(string1.removeLast(-5), 'three');
+      expect(string1.removeLast(6), '');
+    },
+  );
+  test(
+    'Removes the last n characters of the string',
+    () {
+      String? string1 = 'three';
+      expect(string1.maxChars(3), 'thr');
+      expect(string1.maxChars(0), '');
+      expect(string1.maxChars(-5), '');
+      expect(string1.maxChars(6), 'three');
+    },
+  );
+  test(
+    'Reverses slash by providing direction',
+    () {
+      String? string1 = 'C:/Documents/user/test';
+      String? string2 = 'C:\\Documents\\user\\test';
+      expect(string1.reverseSlash(0), 'C:\\Documents\\user\\test');
+      expect(string1.reverseSlash(1), 'C:/Documents/user/test');
+      expect(string1.reverseSlash(-5), 'C:/Documents/user/test');
+      expect(string1.reverseSlash(6), 'C:/Documents/user/test');
+
+      expect(string2.reverseSlash(0), 'C:\\Documents\\user\\test');
+      expect(string2.reverseSlash(1), 'C:/Documents/user/test');
+      expect(string2.reverseSlash(-5), 'C:\\Documents\\user\\test');
+      expect(string2.reverseSlash(6), 'C:\\Documents\\user\\test');
+    },
+  );
+  test(
+    'Returns the character at index',
+    () {
+      String? string1 = 'esentis';
+      expect(string1.charAt(0), 'e');
+      expect(string1.charAt(1), 's');
+      expect(string1.charAt(-5), null);
+      expect(string1.charAt(6), 's');
+      expect(string1.charAt(12), null);
+    },
+  );
+  test(
+    'Appends a specific text to string',
+    () {
+      String? string1 = 'esentis';
+      expect(string1.append(' is trying to be a developer'),
+          'esentis is trying to be a developer');
+    },
+  );
+  test(
+    'Formats the string to price amount',
+    () {
+      String? s1 = '123';
+      expect(s1.toPriceAmount(currencySymbol: '€'), '123,00 €');
+      String? s2 = '123333333';
+      expect(s2.toPriceAmount(), '123.333.333,00');
+      String? s3 = '';
+      expect(s3.toPriceAmount(currencySymbol: '€'), '');
+      String? s4;
+      expect(s4.toPriceAmount(currencySymbol: '€'), null);
+      String? s5 = '044ff1231234566123';
+      expect(s5.toPriceAmount(currencySymbol: '€'), null);
+      String? s6 = '45.225422';
+      expect(s6.toPriceAmount(currencySymbol: '€'), '45,23 €');
+      String? s7 = '45,225422';
+      expect(s7.toPriceAmount(), '45,23');
+      String? s8 = '45,2254,22';
+      expect(s8.toPriceAmount(currencySymbol: '€'), null);
+      String? s9 = '-1245';
+      expect(s9.toPriceAmount(currencySymbol: '€'), '-1.245,00 €');
+      String? s10 = '-01245';
+      expect(s10.toPriceAmount(currencySymbol: '€'), '-1.245,00 €');
     },
   );
 }